fixed tournament bug after refactoring simulationservice location
